+++ src/main/java/ladysnake/sculkhunt/common/block/SculkVeinBlock.java
@@ -44,14 +44,6 @@ public class SculkVeinBlock extends AbstractLichenBlock implements Waterloggable
         builder.add(WATERLOGGED);
     }
 
-    public BlockState getStateForNeighborUpdate(BlockState state, Direction direction, BlockState neighborState, WorldAccess world, BlockPos pos, BlockPos neighborPos) {
-        if (state.get(WATERLOGGED)) {
-            world.getFluidTickScheduler().schedule(pos, Fluids.WATER, Fluids.WATER.getTickRate(world));
-        }
-
-        return super.getStateForNeighborUpdate(state, direction, neighborState, world, pos, neighborPos);
-    }
-
     public FluidState getFluidState(BlockState state) {
         return (Boolean) state.get(WATERLOGGED) ? Fluids.WATER.getStill(false) : super.getFluidState(state);
     }