+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Sculkhunt - Changelog:
 
-### Illuminations 1.1 - 1.17.1
+### Sculkhunt 1.1.1 - 1.18.1
+- Updated to Minecraft 1.18.1
+- Removed the damage ramping up the more players a sculk tracker is near
+- Damage now scales down the more sculk trackers there are
+
+### Sculkhunt 1.1 - 1.17.1
 - Ringing bells now causes random reverberation vibrations in a 30 block radius, allowing you to hide your vibrations to sculk players
 - Added a blacklist and sculkBlacklist command to select certain people that will be guaranteed as survivors at the end of the preparation phase
 - Sculk trackers will now be unable to spawn in a 30 block radius around survivors
@@ -14,6 +19,6 @@
 - Fixed trackers respawning in walls at underground catalysts, therefore causing to go up and being unable to relocate after respawn with the catalyst
 - Increased sculk health for players turned into sculk after the preparation phase to 6 hearts
 - Reduced default catalyst territory radius gamerule to 100 (was 200)
-- 
+
 ### Sculkkhunt 1.0 - 1.17.1
 - Initial release